refactor bao-video structure to facilitate hw and hosted modes
