+++ services/bao-video/Cargo.toml
@@ -13,7 +13,7 @@ log = "0.4.14"
 num-derive = { version = "0.4.2", default-features = false }
 num-traits = { version = "0.2.14", default-features = false }
 cramium-hal = { path = "../../libs/cramium-hal", features = ["std"] }
-cram-hal-service = { path = "../cram-hal-service" }
+cram-hal-service = { path = "../cram-hal-service", optional = true }
 ux-api = { path = "../../libs/ux-api" }
 
 # QR decoding
@@ -24,6 +24,12 @@ rqrr = "0.8.0"
 locales = { path = "../../locales" }
 
 [features]
+# This crate is baosec specific; these flags switch between hosted mode emulation
+# (for UI development) and actual hardware targets. In UI development all camera data
+# is "faked"
+board-baosec = ["cram-hal-service"]
+hosted-baosec = []
+
 cramium-soc = ["utralib/cramium-soc"]
 decongest-udma = []
 default = ["decongest-udma"]
