+++ services/bao-video/src/api.rs
@@ -1,0 +1,5 @@
+#[derive(num_derive::FromPrimitive, num_derive::ToPrimitive, Debug)]
+pub enum Opcode {
+    CamIrq,
+    InvalidCall,
+}
