+++ services/bao-video/src/lib.rs
@@ -1,0 +1,2 @@
+mod api;
+pub use api::*;
